requirements.txt
langchain==0.0.237
python-dotenv==1.0.0
typing-inspect==0.8.0
typing_extensions==4.5.0
pydantic==1.10.11
unstructured[local-inference]
layoutparser[layoutmodels,tesseract]
openai